Replace shifting by FP_BASE with mul/div by FP_FACTOR
The left and right shift operators in C, not being typical arithmetic
operators, have some edge cases which make them awkward to use in a portable
way. For example, left shifting a negative number is undefined behaviour,
and right shifting a negative number is implementation-defined.

They also do not invoke the usual arithmetic conversions which occur with
other operators. If we left shift an int, the result will remain an int,
which might not be able to fit the intended value. But if we multiply by
an appriopriately-typed power of two, the int will be cast to the correct
type. This helps prevent portability errors.