Reveal hidden doors and levers when moving into them
This circumvents an issue introduced with the new multi-turn search mechanic:
if the player is trapped up against an obvious secret door by an enemy, they
must take 5 turns worth of damage before they can go through the door. This
is frustrating for the player, who has already "found" the door.