monsterDetails: Show player-to-monster combat info for allies
Now that the wand of plenty changes max HP, it was difficult to identify
how many times an ally had been cloned, e.g. if you wanted to spread out
the health pool as much as possible. This change detaches the statement of a
monster's allegiance from the "blue" combat info and puts it in its own section
before. This is also generally useful information for inspecting your allies.