Record health changes at the start of playerTurnEnded
Previously it was at the end, and because monsters are moved in the body of
this function, that meant that health changes from them sometimes weren't
shown. However this is still not perfect, as we forget any health changes
that occur as a result of player actions. (I considered this less important
than seeing how much damage was being taken.)