Moved charm change to Bug fixes in Changelog
The charm changes in 1.7.5-flend1 restore all charms to v1.7.4 behaviour
where the charms were affected by bugs introduced during the change to
fixed point in v1.7.5. There were no intended charm changes according to
the v1.7.5 changelog and it was clear the changes were due to bugs (and
unbalanced the game in many cases). The charms now behave as in v1.7.4
although there are some off-by-1 changes that are not obviously bugs,
just side effects of the new maths.