Assign a dummy date to a file before reading the real one
I tried to keep the same spirit as the original code, where "\0" was
assigned to a string, probably to avoid random data and keep the
recordings list tidy in case of error while reading the real date.

It seems assigning {0} to a struct tm (propagating to all its members)
is the closest thing I could do to replicate this behaviour, as I
don't think it's possible to assign "empty" values to a struct tm.

I also fixed indentation & formatting errors.