Fix sprintf warnings
Those warnings worded differently on different compilers, but they all
boil down to the fact that the same source and target is used in a
sprintf. This commit makes sure, that these are being written via an
intermediate buffer. It also might have fix a bug discussed on discord
as Tom believes that the bug is a manifestation of this issue