Check for NO_DIRECTION in pathTowardCreature
Previously, if nextStep returned NO_DIRECTION (-1) an invalid access would
occur, causing an out-of-sync error. (#70)

This also slightly changes the logic. Previously, if moveMonsterPassivelyTowards
failed, a random step would be tried. Now a random step is only tried if
nextStep fails. I think this should be almost the same, since nextStep checks
whether it can pass any adjacent monsters, which seems to be why the
moveMonsterPassivelyTowards return value was checked.