Fix time arithmetic (curses)
`uint32_t` is not large enough to contain the UTC epoch in milliseconds,
so the output always wrap around. It is not a problem in practice:
at worst the game will occasionally wait 0 ms instead of 34 ms.
But switching to `uint64_t` gets rid of wraparound behavior here
and any runtime debug warning associated with it.