Include the seed and depth into save and recording file names
Recordings also indicate if the player won, died or quit.

This patch is forward-compatible with 64-bit seeds.
If the seed is very long it gets shortened to 11 characters max,
to keep a reasonable file name length.

Examples:

- `Saved #1817262093 at depth 37 (wizard).broguesave`
- `#148...15217 Quit at depth 12 (easy).broguerec`
- `#3611284904 Died at depth 7 (2).broguerec`
- `#2852666709 Died at depth 19.broguerec`
- `#3898787308 Mastered the dungeons.broguerec`
- `#1631136788 Escaped the dungeons.broguerec`

With WebBrogue, file names must fit into 30 bytes, so
we'll only show the seed, for example:

- `#1234567890.broguerec`
- `#1234567890_2.broguerec`
- `#1234567890_3.broguerec`
- `#123...45678.broguerec` (64-bit seeds)
- `#123...45678_2.broguerec`