Fix invalid pointers to freed memory
When a monster is demoted, we are assigning a new (or no) leader to
the monsters on the same level but not to those on other levels.
When the monster dies, its memory is freed. Its former followers
now hold pointers to freed (and possibly reassigned) memory. This
wasn't usually causing a crash, but it was making the game state
dependent on `malloc`'s memory management, which isn't good.

This patch checks all levels, to correctly clean up all `leader`
pointers.

One of the games on WebRogue's top list was affected by it, resulting
in OOS. Possibly many other games that go OOS have the same root
cause.

Submitting the patch to `master`, not `release`, because it breaks
recordings made with Brogue CE 1.9, which unintentionally *relies*
on pointers pointing to the same invalid memory addresses during
playback as they did during recording.