Fix out-of-sync error when escaping from a net or web
When the player escapes from being stuck, a movement was recorded without
returning from playerMoves. This was a problem because confirmation prompts
(move into chasms, attack an acid mound etc.) were processed later, potentially
aborting the move and recording an event that didn't actually happen. The
STATUS_STUCK code is moved (in a backwards-compatible way) to after the
confirmation prompts, ensuring the turn will always end after recording.

Additionally in the new code, the moveEntrancedMonsters() call is moved into
only the ending-turn branch to avoid it being called twice when the player is
unstuck.

Co-authored-by: Sanat Mishra <sanatmishra2222@gmail.com>
Co-authored-by: Antony Boucher <69065091+AntonyBoucher@users.noreply.github.com>