Fix a crash when moving entranced monsters (#155)
moveEntrancedMonsters didn't account for a monster dying in the middle of the
turn. If an entranced monster moves and the next monster in the chain dies, the
rest of the function iterated across the rest of the dead monsters from that
turn, potentially moving an already dead monster and causing crashes.