Scalable window
The window is a mosaic of tiles which are either `N` or `N+1` pixels
wide, and either `M` or `M+1` pixels tall. The large PNG is loaded
once into `Tiles[0]`. When the window size changes, `Tiles[1..4]` are
regenerated with copies of it, scaled down to the four possible
combinations of tile width and height. Each tile gets scaled down
separately, and only when first needed -- not upfront.

The downscaling is performed in linear color space. The tile is
logically split into 16 parts (sliced by 3 vertical lines and 3
horizontal ones), so that:

- the four corners can be positioned to preserve the tile's aspect
  ratio, possibly filling the blank spaces at the top and bottom;
- the top, bottom, left, right, and center areas can be independently
  aligned with output pixels to improve sharpness.

Some tiles, like walls and doors, always stretch to fill the space
regardless of aspect ratio. Other tiles must preserve aspect ratio,
but we allow up to 20% stretch for graphic tiles and 40% for text
tiles.

For text tiles (letters, digits and punctuation signs), the
characters' x-height and baseline (corresponding to the top and bottom
of "x") are pixel-aligned regardless of outlines. This ensures that all
letters neatly line up without jumping up and down.

Pixel alignment is discovered by brute-force, trying all possible
alignments in 0.1 pixel increment (up to 0.5 pixel). Because this is
a time-consuming process, it is done only once and saved to disk
(`tiles.bin`, committed to git along `tiles.png`). If you ever modify
`tiles.png`, you should regenerate `tiles.bin`: simply delete the
existing file then start Brogue.

Wall tops are diagonal sine waves, approximately 4 pixels apart. If the
floor tile is completely blank, we fill it with dust-like noise
(isolated dots).