Show git commit and branch name in dev builds (#199)
Call a build a release build either if RELEASE is YES, or if it is built
in CI from the gha-build branch.  Call others development builds.

For development builds built by github actions or locally with a git
checkout, append the git commit and branch name onto the version.  This
makes the extra fields visible in e.g. the title screen and --version
output.  If built locally without git, just append "dev" instead.