Center the glyph preview window and allow closing it
When Brogue needs to regenerate tiles.bin
a small preview window opens to show its progress.
This commit centers the window so that it is
where the game window would have opened / will open.

The preview window was not easy to close because
we ignored all events until completion. On some system
this may even prevent the preview window from showing
on screen.