Additional I/O checks
Verify that tiles.png exists before trying to load it.
We cannot rely on IMG_Load() because it uses special
tricks on macOS to find the file relatively to either
the current directory or the install directory. So,
when running "bin/brogue" (without MAC_APP=YES) we would
load the PNG just fine but fail to find tiles.bin next to it
because we load tiles.bin with fopen() which doesn't
employ such tricks, causing Brogue to needlessly
optimize the tiles instead of starting the game.

When writing tiles.bin back, also verify that the file
could be opened in write mode.