Collapse repeated archived messages
..and display them with a count, capping out at 100.

A message will collapse if it shares the same content  (ignoring the
count) under one of two related but separate conditions.

1. Uninterrupted runs of repeated non-combat messages will collapse
regardless of what turn they occurred.

2. Separately, any repeated combat messages arriving on the same player
turn will collapse, regardless of arrival order.

Combat and non-combat messages never collapse together.

The implementation involves bulk-redrawing archived messages in the
displayed messages area.  This is encapsulated in the new
displayRecentMessages() function.

We put it to use to redraw the recent archive after canceling a save,
which in my opinion gives a better visual experience.