Portable patchVersion checks
Right now, the `master` branch still carries the version number `1.9.4`. This is problematic because any build made off master will show "1.9.4" in the main menu, which is misleading. We cannot safely change it to `1.10.0` though, because doing so will disable features introduced since 1.9.1, and restore bugs fixed since 1.9.1, due to the dozens of `patchVersion` checks which have not yet been removed from the code merged into `master`.

Removing the `patchVersion` checks in master is simple (turn the comparisons into either `true` or `false`, then simplify the code) but it causes annoying merge conflicts which we would rather not deal with on an ongoing basis.

This patch offers a solution, so that we can happily merge `release` into `master` without having to worry about refactoring the version checks or dealing with the regressions that leaving them in would cause. If we want to remove the checks for the 1.10 release, we can do it all the same, but there are no consequences if we don't (apart from some dead code).

Instead of writing
    `rogue.patchVersion >= 4`
you would write
    `COMPARE_TO_VERSION(1,9,4) >= 0`
and instead of
    `rogue.patchVersion < 4`
this
    `COMPARE_TO_VERSION(1,9,4) < 0`

The new `COMPARE_TO_VERSION` macro takes the Major and Minor versions into consideration, not just the Patch version.

Major and Minor are tracked with new defines, `BROGUE_MAJOR` and `BROGUE_MINOR` which will be `1` and `9` in `release` branch, `1` and `10` in `master`. There is also `BROGUE_PATCH` now (but the `COMPARE_TO_VERSION` uses `patchVersion`, of course).