Press w to hot-swap equipment
Whenever a piece of equipment displaces another when being equipped, it
is "swapped in" and the other is "swapped out".  Pressing "w" then
exchanges these two as a player action.  In this way players can
"juggle" rings, weapons or armor with a single key stroke.

If either of the swapped in or out items leaves the pack, the swapped
in/out status of both is cleared.  If the swapped in item is cursed,
swapping fails with an appropriate message, but the status is not
cleared.  There is no visual feedback (in the inventory screen or
elsewhere) of which items are swapped in or out.

Closes #220