Allow seeking to the last turn of a recording
Playback seeking was artificially limited to the penultimate turn,
because the `gameOver()` / `victory()` could not display the last turn
in fast-forward mode. Fixing the latter now allows seeking to the
last turn.

For convenience, we'll also allow entering a large turn number,
e.g "99999" (easy to type!) which is now treated the same as
seeking the end of the recording (i.e last turn).