Record the keystroke up-front but allow cancelling (#305)
Previously, playerMoves() called handleWhipAttacks() before recording the
keystroke. handleWhipAttacks() calls zap(), which calls updateBolt(), which
calls gameOver(), which calls flushBufferToFile() then saveRecording(). At that
point, it was too late for the last keystroke to make it into the recording
file. We ended up with a recording that was missing the last keystroke, and
playback gave an OOS error when it tried to read more events at the end of the
recording (since the player is not dead yet).