Ignore fallen items in earlier simulated turns
Items that fall into a level should not be affected
by the environment (burnt, moved, etc)
for more turns than they spent in that level.
By tracking the turn when an item falls, we can
ignore it in earlier simulated turns,
even though it is already placed on the floor.

This solves the issue that if you fall into a level
at the same time or soon after an item, the
item may have floated very far in a pool, or burned
entirely in lava. Now it will have floated/burned
for just the right number of turns.