Don't print equip messages when throwing (#329)
Throwing (and enchanting, falling victim to acid mounds, or armor of
burden..) implicitly equips an item as part of its implementation.
These calls can now produce undesired messages.

These "hidden" equip actions set the force parameter to true, so be
quiet when this flag is set.  The strengthCheck function can also
produce messages this way, give it a flag to suppress messages as well.