Cap exposeTileToFire calls per tile per update (#344)
In most games, tiles are normally exposed to fire up to 5 times per
environment update. In rare circumstances (shooting a firebolt into
a fire?) it can go up to 6 times. But it never goes higher than
that... except when a bug causes an infinite loop involving
`promoteTile()` and `exposeTileToFire()`, in which case the same
tile can be exposed to fire over a thousand times, until Brogue
runs out of stack space and crashes.

This commit limits `exposeTileToFire()` to 12 calls per tile per
environment update. This fixes #341, without causing OOS on games
unaffected by the bug. Tested on 10 victories and 10 ascensions
from WebBrogue.