Fix allies casting harmful bolts at the player
The condition that checked whether the creature on a bolt's trajectory
is on the same side as the bolt's caster didn't work when the caster
is an ally of the player. In that case it gave a positive bonus for hitting
the creature, instead of a negative one. The issue was that if the bolt
was not cast by the player, then it was assumed to be cast by an enemy
of the player.