Auto-kink unknown bolts like a throw
Until a wand or staff is identified, we need to hide whether it targets
allies or enemies. We are achieving it by disabling the auto-kinking,
but it would be more convenient for the user that we treat it like a
throw and do our best effort to reach the target cell.