Fix build errors when compiling with TERMINAL=YES or WEBBROGUE=YES
Commit 5d2b549414d4c0c4d36bdcc952deb1440b1ff6f1 appears to have
introduced a find-and-replace error, incorrectly substituting the xLoc
and yLoc parameters of curses_plotChar and web_plotChar with loc.x and
loc.y, respectively. This changeset reverts these parameter names in
order to un-break the build when compiling with TERMINAL=YES or
WEBBROGUE=YES.

Additionally this changeset fixes a format-security error in
web-platform.c which was also causing a build failure.