Fix build errors with TERMINAL=YES or WEBBROGUE=YES (#382)
Commit 9a83649c081bedc7af96a3cc54a49d163f034a64 appears to have
introduced a find-and-replace error, incorrectly substituting the xLoc
and yLoc parameters of curses_plotChar and web_plotChar with loc.x and
loc.y, respectively. This changeset reverts these parameter names in
order to un-break the build when compiling with TERMINAL=YES or
WEBBROGUE=YES.

Additionally this changeset fixes a format-security error in
web-platform.c which was also causing a build failure.