Fix rounding errors when estimating value of unidentified armor (#390)
When the armor rating is displayed in the info dialog, the previous
process of estimating the armor rating is something like this:

    estimated = <armor_rating>                                    // => 50
    estimated = estimated + (10 * strength_modifier / FP_FACTOR)  // => 50 - 25 = 25
    estimated = estimated / 10                                    // => 25 / 10 = 2

But when it's equipped, the process ( IO.c:estimatedArmorValue ) is more
like:

    estimated = <average_armor_rating> / 10                       // => 5
    estimated = estimated + (strength_modifier / FP_FACTOR)       // => 5 + 2 = 3

Note, `10 * strength_modifier / FP_FACTOR` is 25 but
`strength_modifier / FP_FACTOR` is 2.

This commit modifies the process to estimate the value of unidentified,
unequipped armor to be exactly the same as the process to estimate
unidentified but equipped armor. The average value of armor is used
(instead of the exact value), and the rounding issues are fixed.