Clarify description of negation charm
The previous description was easy to misread as "the charm will negate all magical effects on (the creatures in your field of view) and (the items on the ground up to 7 spaces away)", which makes it sound like it affects all creatures in your field of view. This rewords the description to hopefully avoid this issue.