Fix crash when an explosive monster dies from poison or burning damage
This crash happened with seed 500500500; if you get the staff of poison from depth 7 and use it on the explosive goblin on depth 11 twice, the game crashes when the goblin dies. It can also happen with burning damage.

This fixes the crash by not removing monsters from `monsters` or `dormantMonsters` until `emptyGraveyard`, as well as modifying `iterateCreatures` to skip dying monsters. This approach should hopefully fix not only this crash, but also any similar crashes that haven't been seen yet.