Specify the numerical benefits of Awareness rings
Enchantment affects the strength of your passive searches as well as the
distance to which those searches are effective.

Findings that were used to create this tooltip:
Passive search radius is 2 without a ring. The radius increases by 2 for
each point of positive Awareness possessed by the player. There is a 10%
chance to detect traps at the outer edge of that radius. Traps are 10%
more likely to be noticed for each space closer to the player.

Search strength is reduced by 1/3 when it comes to secret doors, but
adding numbers for that to the tooltip makes it extremely wordy.