Nerf the speed weapon runic, and adjust 2 others
Multiple players have been posting extremely successful games with
weapons of speed, lately. These weapons trigger much more reliably than
any other weapon, and the whole dungeon stands still while the player
chops down a monster, which can, like quietus, result in few or no
counterattacks.

This runic can yield a ton of juice reaping/transference, however, where
quietus does not. Is it as powerful as the Confusion runic? Almost
certainly, if not more powerful. Speed compares quite favorably with
slowing and paralysis, which are only able to deal with one monster at a
time.

But let's try dropping this from 16 to 11 before we go any further. And
while we're at it, Multiplicity is *definitely* more powerful than
Force, so nudge that down, and Slowing is *definitely* the weakest
runic, so move it to the very-high-proc-chance tier.