Improve exit game text and remove load & replay commands (#475)
"Quit without saving" is now "Quit and abandon game"; "Suspend game and quit" is
now "Save and exit". These commands are now next to each other in the menu.

Also remove the commands for opening saved games and recordings from within a
game, as they could be used for save scumming and don't seem to have much
purpose otherwise; both of those actions can now only be done from the main menu
or from a recording.