Fix mirrored totem beckoning sometimes instakilling the player (#498)
This fixes the issue by reversing the bolt direction for the blinking triggered
by beckoning, by swapping the bolt origin and target and then reversing the
resulting list.

This also makes the same adjustment to the behavior of wands of beckoning. This
seems to make the wand's behavior more closely match what you'd expect from the
path displayed when targeting it. For instance, it makes it easier to target
enemies that are behind other enemies.