Fix unused function and missing prototype warnings
Some of these warnings were directly introduced in #593; others are due to other PRs that were merged between when I made that PR and when it was merged.

Some of these are fixed with new `#define`s; this also makes `displayLoops` and `displayChokeMap` static and always runs them in wizard mode when '~' is pressed, along with `displayMachines` and `displayWaypoints`. This also improves those functions a bit, so it's easier to tell which output is from which function.